backend/requirements.txt
# backend/requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
python-socketio==5.5.2
# Add other dependencies as needed